Combine separate files into regions in module file
